feat(setup): changed theme color definitions
