+++ src/components/BaseInput/base-input.styles.ts
@@ -10,7 +10,7 @@ export const Wrapper = styled.View`
 export const InputLabel = styled.Text<LabelProps>`
   margin-bottom: ${props => props.theme.spacings.xs};
 
-  color: ${props => props.theme.colors.text.main};
+  color: ${props => props.theme.colors.text.default.main};
   font-size: ${props => props.theme.fontSizes.lg};
   text-transform: ${props => (props.preventCapitalize ? 'none' : 'uppercase')};
   font-weight: bold;
@@ -21,7 +21,7 @@ export const Input = styled.TextInput`
   padding: 0 ${props => props.theme.spacings.md};
 
   font-size: ${props => props.theme.fontSizes.lg};
-  border-color: ${props => props.theme.colors.primary.main};
+  border-color: ${props => props.theme.colors.primary.default.main};
   border-width: ${props => props.theme.borderWidth.sm};
   border-radius: ${props => props.theme.borderRadius.md};
 `;
@@ -29,7 +29,7 @@ export const Input = styled.TextInput`
 export const ErrorMessage = styled.Text`
   margin-top: ${props => props.theme.spacings.xs};
 
-  color: ${props => props.theme.colors.danger.main};
+  color: ${props => props.theme.colors.danger.default.main};
   font-size: ${props => props.theme.fontSizes.md};
   font-weight: 500;
 `;
