+++ src/screens/SignIn/sign-in.styles.ts
@@ -14,7 +14,7 @@ export const Wrapper = styled(ScreenWrapper)`
 export const ConfirmButton = styled(BaseButton)``;
 
 export const LoginIcon = styled(MaterialCommunityIcon).attrs(props => ({
-  color: props.theme.colors.primary.contrast,
+  color: props.theme.colors.primary.default.contrast,
   name: 'login-variant',
   size: 24,
 }))``;
